Time: 47 ms (74.21%) | Memory: 151.7 MB (73.97%) - LeetSync

+++ 2403-count-unreachable-pairs-of-nodes-in-an-undirected-graph/count-unreachable-pairs-of-nodes-in-an-undirected-graph.cpp
@@ -37,12 +37,12 @@ public:
         }
 
         long long result = 0;
-        long long remainingNodes = n;
+        int remainingNodes = n;
 
         for(auto &it: mpp){
             remainingNodes -= it.second;
 
-            result += it.second*remainingNodes;
+            result += (it.second*(long long)remainingNodes);
         }
 
         return result;
